# GH-2442 value comparator benchmarks and equals