--- a/core/queryalgebra/evaluation/pom.xml
+++ b/core/queryalgebra/evaluation/pom.xml
@@ -25,7 +25,7 @@
 			<artifactId>rdf4j-query</artifactId>
 			<version>${project.version}</version>
 		</dependency>
-		<!-- FIXME: Move SPARQLFederatedService out to another module to avoid the following 
+		<!-- FIXME: Move SPARQLFederatedService out to another module to avoid the following
 			dependency -->
 		<dependency>
 			<groupId>${project.groupId}</groupId>
@@ -53,6 +53,18 @@
 		<dependency>
 			<groupId>com.google.guava</groupId>
 			<artifactId>guava</artifactId>
+		</dependency>
+		<dependency>
+			<groupId>org.openjdk.jmh</groupId>
+			<artifactId>jmh-core</artifactId>
+			<version>${jmhVersion}</version>
+			<scope>test</scope>
+		</dependency>
+		<dependency>
+			<groupId>org.openjdk.jmh</groupId>
+			<artifactId>jmh-generator-annprocess</artifactId>
+			<version>${jmhVersion}</version>
+			<scope>test</scope>
 		</dependency>
 	</dependencies>
 	<build>