--- a/core/sail/shacl/pom.xml
+++ b/core/sail/shacl/pom.xml
@@ -57,7 +57,6 @@
 			<version>${jmhVersion}</version>
 			<scope>test</scope>
 		</dependency>
-		<!-- https://mvnrepository.com/artifact/org.openjdk.jmh/jmh-generator-annprocess -->
 		<dependency>
 			<groupId>org.openjdk.jmh</groupId>
 			<artifactId>jmh-generator-annprocess</artifactId>