… core/queryalgebra/evaluation/src/test/resources/bsbm-100.ttl