@hmottestad 2020-08-17
# Benchmark results

## Before
```
Benchmark                                       Mode  Cnt    Score    Error  Units
ValueComparatorBenchmark.sortManyDeepEquals     avgt    5  452.960 ± 29.174  ms/op
ValueComparatorBenchmark.sortManyPointerEquals  avgt    5  157.041 ±  6.604  ms/op
ValueComparatorBenchmark.sortObjects            avgt    5  250.146 ±  3.246  ms/op
ValueComparatorBenchmark.sortPredicates         avgt    5    0.107 ±  0.011  ms/op
ValueComparatorBenchmark.sortSubjects           avgt    5  104.293 ±  0.666  ms/op
```

## After
```
Benchmark                                       Mode  Cnt    Score    Error  Units
ValueComparatorBenchmark.sortManyDeepEquals     avgt   10  485.574 ± 18.843  ms/op
ValueComparatorBenchmark.sortManyPointerEquals  avgt   10  132.606 ±  0.654  ms/op
ValueComparatorBenchmark.sortObjects            avgt   10  207.202 ±  3.310  ms/op
ValueComparatorBenchmark.sortPredicates         avgt   10    0.093 ±  0.001  ms/op
ValueComparatorBenchmark.sortSubjects           avgt   10   79.736 ±  0.494  ms/op
```