@hmottestad 2020-08-17
The goal here was to simplify the comparison code. Essentially, the complexity for .equals(...) is similar to doing a comparison. This means that we can remove the .equals and just rely on the comparison code. The .equals(...) is inside the call to ObjectUtil.nullEquals(...). We still check for simple pointer equality.