@hmottestad 2020-08-17
Seems to give a decent ~20% performance improvement without any significant performance hits.